add generic banner component
