--- src/app/(main)/apoyoalcuidado/adapters/index.js
@@ -1,1 +1,0 @@
-export * from './resourcesCardsAdapter'
