--- src/app/(main)/apoyoalcuidado/components/Tabs/components/index.js
@@ -1,1 +1,0 @@
-export { default as TabCards } from './Cards/TabCards'
