--- src/app/(main)/apoyoalcuidado/components/index.js
@@ -1,3 +1,0 @@
-export { default as ContainerView } from './ContainerView/ContainerView'
-export { default as ResourcesTopView } from './ResourcesTopView/ResourcesTopView'
-export { default as Tabs } from './Tabs/Tabs'
