--- src/app/(main)/apoyoalcuidado/hooks/index.js
@@ -1,1 +1,0 @@
-export { default as useResources } from './useResources'
