--- src/app/(main)/apoyoalcuidado/hooks/useResources.jsx
@@ -1,8 +1,0 @@
-import { useContext } from 'react'
-import ResourcesContext from '../context/ResourcesProvider'
-
-const useResources = () => {
-  return useContext(ResourcesContext)
-}
-
-export default useResources
