--- src/app/(main)/apoyoalcuidado/page.jsx
@@ -1,13 +1,0 @@
-'use client'
-import { ContainerView } from './components'
-import { ResourcesProvider } from './context/ResourcesProvider'
-
-function RecursosDeApoyo() {
-  return (
-    <ResourcesProvider>
-      <ContainerView />
-    </ResourcesProvider>
-  )
-}
-
-export default RecursosDeApoyo
