+++ src/app/(main)/apoyoalcuidador/adapters/index.js
@@ -1,0 +1,1 @@
+export * from './resourcesCardsAdapter'
