+++ src/app/(main)/apoyoalcuidador/components/Tabs/components/index.js
@@ -1,0 +1,1 @@
+export { default as TabCards } from './Cards/TabCards'
