+++ src/app/(main)/apoyoalcuidador/components/index.js
@@ -1,0 +1,3 @@
+export { default as ContainerView } from './ContainerView/ContainerView'
+export { default as ResourcesTopView } from './ResourcesTopView/ResourcesTopView'
+export { default as Tabs } from './Tabs/Tabs'
