+++ src/app/(main)/apoyoalcuidador/hooks/index.js
@@ -1,0 +1,1 @@
+export { default as useResources } from './useResources'
