+++ src/app/(main)/apoyoalcuidador/hooks/useResources.jsx
@@ -1,0 +1,8 @@
+import { useContext } from 'react'
+import ResourcesContext from '../context/ResourcesProvider'
+
+const useResources = () => {
+  return useContext(ResourcesContext)
+}
+
+export default useResources
