+++ src/app/(main)/components/GenericBanner/GenericBanner.module.css
@@ -1,0 +1,76 @@
+.banner{
+    display: none;
+    background-color: var(--clr-secondary-1);
+}
+
+@media (min-width: 576px) {
+    .banner{
+        display: grid;
+        grid-template-columns: repeat(12, 1fr);
+        gap: 5px;
+        margin: 1rem;
+        border-radius: 20px;
+        margin-top: 1.5rem;
+    }
+    .bannerImg{
+        grid-column: 1 / 8;
+        aspect-ratio: 16/9;
+        width: 100%;
+        height: auto;
+    }
+    .bannerTextBox{
+        grid-column: 8 / -1;
+        display: grid;
+        place-items: center;
+    }
+    .textBox{
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        gap: 10px;
+
+        text-align: center;
+        max-width: 274px;
+    }
+    .bannerTitle{
+        font-family: var(--ff-one);
+        font-weight: 500;
+        color: var(--clr-secondary-2);
+        font-size: clamp(1.4rem, 4vw, 2.4rem );
+    }
+    .titleStrong{
+        font-family: var(--ff-two);
+        font-style: italic;
+        font-weight: 400;
+        color: var(--clr-primary-2);
+    }
+    .bannerMessage{
+        font-family: var(--ff-one);
+
+        font-weight: 300;
+        font-size: clamp(0.8rem, 2vw, 1.2rem);
+    }
+
+} /* pantalla mediana */
+@media (min-width: 992px) {
+    .banner{
+        margin-top: 2rem;
+    }
+    .textBox{
+        gap: 20px;
+        max-width: 320px;
+    }
+    .bannerTitle{
+        font-size: 2.5rem;
+    }
+    .titleStrong{
+        font-size: 1.1em;
+
+    }
+    .bannerMessage{
+        font-size: 1.4rem;
+    }
+
+} /* pantalla grande */
+@media (min-width: 1400px) {
+} /* pantalla extra grande */
